requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
gunicorn==23.0.0
sqlalchemy==2.0.36
alembic==1.14.0
jinja2==3.1.4
python-multipart==0.0.12
pandas==2.2.2
openpyxl==3.1.5
qrcode==7.4.2
pillow==10.4.0
httpx==0.28.1
psycopg[binary]==3.2.3
PyMySQL==1.1.1
paho-mqtt==1.6.1
pytest==8.3.5
